Implement enumeration of all services
Instead of searching for specific UUIDs, search for all primary services.
If a desired service is found, search for its characteristics from the
current handle to the last handle associated with the service. When all
desired characteristics of the service are found, continue searching for
all primary services, with the last handle of the current service as
beginning.

Signed-off-by: Florian Limberger <florian.limberger@inovex.de>